server/requirements.txt
flake8===3.6.0
Flask==1.0.2
flask-cors==3.0.7
flask-migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
